requirements.txt
colorama==0.3.7
coreapi==2.1.1
coverage==4.3.4
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
django-guardian==1.4.6
django-rest-swagger==2.1.1
djangorestframework==3.5.3
itypes==1.1.0
Markdown==2.6.7
nose==1.3.7
openapi-codec==1.2.1
psycopg2==2.6.2
Pygments==2.1.3
python-termstyle==0.1.10
rednose==1.2.1
requests==2.12.5
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
